requirements.txt
Pillow>=10.0.0
pillow_heif>=0.13.0
arrow>=1.2.3